requirements.txt
requests>=2.25.1
python-dateutil>=2.8.1
tabulate>=0.8.9
